refactor: 클래스 역할에 맞도록 S3PoolTask를 AsyncConfig클래스로 이동
